==> 01.2_KaggleCatcher.ipynb <==
# 1. Extracting Kaggle datasets

Daniel Ruiz, MSc in Data Science and Business Analytics (DSBA), Bocconi University

Reference codes (alphabetically):
- https://www.kaggle.com/augustop/portuguese-tweets-for-sentiment-analysis
- https://www.kaggle.com/kazanova/sentiment140

# import packages

import pandas as pd

## English

company= ['EnglishTweets','@ENGLISH_TWEETS','#ENGLISH_TWEETS']
df = pd.read_csv('Dataset_Kaggle\\Kaggle_English\\'+company[0]+'.csv',
                 header=None,index_col=None,encoding='ANSI')

df.columns = ['sentiment_pos','id','datetime','query','user_name','text']
df['user_followers']=1000
df = df[['text','id','datetime','user_name','user_followers','sentiment_pos','query']]
df.sentiment_pos = df.sentiment_pos//4

# add columns -> the values are only so the lines are not excluded

df.to_csv('Dataset_Twitter\\us_EnglishTweets.csv',index=None,header=False,encoding='utf-8')

## Portuguese

# opens file, 'converts' it to our framework
df = pd.read_csv('Dataset_Kaggle\\Kaggle_Portuguese\\TrainingDatasets\\Train500.csv',sep=';')
    
# add columns -> the values are only so the lines are not excluded
df['user_name'] = '@PORTUGUESE_TWEETS'
df['user_followers'] = '1000'

# reorder, rename columns
df = df[['tweet_text','id','tweet_date','user_name','user_followers','sentiment','query_used']]
df.columns = [['text','id','datetime','user_name','user_followers','sentiment_pos','query_used']]

# save
df.to_csv('Dataset_Twitter\\br_PortugueseTweets.csv',index=None,header=False)
