==> 01.3_FileMerger.ipynb <==
# 1. Merge CSVs of the Same Company

Daniel Ruiz, MSc in Data Science and Business Analytics (DSBA), Bocconi University

Reference codes (alphabetically):

## 1.1. Running the code

# import packages
import pandas as pd

# merge files
folder = 'Dataset_Twitter/'

file1 = pd.read_csv(folder+'us_wsj.csv')
file1.columns = ['text','id','date','user','followers']
file1 = file1.drop_duplicates('id')
print(file1.shape)

file2 = pd.read_csv(folder+'us_wsj_01.csv')
file2.columns = ['text','id','date','user','followers']
file2 = file2.drop_duplicates('id')
print(file2.shape)

merged = pd.concat([file1, file2])
print(merged.shape)

merged.columns = ['text','id','date','user','followers']
#merged = merged.drop_duplicates('id')
print(merged.shape)

merged = merged[merged.id!=1249859590105239552]

merged.to_csv(folder+'us_wsj_merged.csv', index=False, header=False)
