==> 01.4_StockTracker_MSN_15.ipynb <==
# 4. Getting financial data

Daniel Ruiz, MSc in Data Science and Business Analytics (DSBA), Bocconi University

Reference codes:

# import packages
import csv
import time

# chromedriver (driver and chrome should be of the same version)
# https://chromedriver.chromium.org/downloads
from selenium import webdriver

# parser
from bs4 import BeautifulSoup

def convert_time_15(date):

    clock, half, dayofweek, month, day = date.split(' ')
    hh, mm = clock.split(':')

    hh = int(hh)
    if half=='PM' and hh!= 12:
        hh+=12
    hh = str(hh)
    if len(hh)==1:
        hh='0'+hh

    if len(day)==1:
        day="0"+day

    str_to_num = {'Jan':'01','Feb':'02','Mar':'03',
                  'Apr':'04','May':'05','Jun':'06',
                  'Jul':'07','Aug':'08','Sep':'09',
                  'Oct':'10','Nov':'11','Dec':'12'}

    return "2020-"+str_to_num[month]+"-"+day+" "+hh+":"+mm+":00"

def StockTracker_15(sup, driver = None):
    
    # sup format = [folder name, url]
    # date format = "2020-01-17"
    
    # print stock
    print(sup[0])
    
    # open chrome driven
    if driver == None:
        driver = webdriver.Chrome(executable_path=r'C:\Users\drr19\Desktop\chromedriver.exe')
    
    # open page, wait loading, scroll dow, and get content
    driver.get(sup[1])
    time.sleep(5)
    driver.find_element_by_xpath("//*[@id='heropanechart-mode-selector']/div/div/div[2]/button").click()
    time.sleep(10)
    driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")
    content = driver.page_source
    
    # parsing the content

    soup = BeautifulSoup(content, "lxml")

    # extracting the variables
    minutes = soup.find("div", {"class":"historicalpricestable"})
    
    timms = minutes.find_all("li", {"class":"rowvalue date"})
    highs = minutes.find_all("li", {"class":"rowvalue high"})
    lowws = minutes.find_all("li", {"class":"rowvalue low"})
    opees = minutes.find_all("li", {"class":"rowvalue open"})
    closs = minutes.find_all("li", {"class":"rowvalue close"})
    volus = minutes.find_all("li", {"class":"rowvalue volume"})

    # converting to list
    
    datas = []

    for timm, high, loww, opee, clos, volu in zip(timms,highs,lowws,opees,closs,volus):
        datas.append([convert_time_15(timm.text),
                      high.text,
                      loww.text,
                      opee.text,
                      clos.text,
                      volu.text.replace(",",".")])

    # saving as csv
        
    for data in datas:
        with open('Dataset_MSN_15\\'+sup[0]+'.csv', "a", encoding="utf-8") as csvFile:
            
            writer = csv.writer(csvFile)
            writer.writerow(data)    

my_companies = [['br_embraer','https://www.msn.com/en-us/money/stockdetails/history/bsp-embr3/fi-apmhnm'], #0
                ['br_americanas','https://www.msn.com/en-us/money/stockdetails/history/bsp-lame4/fi-apmupr'], #1
                ['br_pontofrio','https://www.msn.com/en-us/money/stockdetails/history/bsp-vvar3/fi-apnla2'], #2
                ['br_petrobras','https://www.msn.com/en-us/money/stockdetails/history/bsp-petr4/fi-apn4gh'], #3
                ['br_bradesco','https://www.msn.com/en-us/money/stockdetails/history/bsp-bbdc4/fi-aplym7'], #4
                ['br_itau','https://www.msn.com/en-us/money/stockdetails/history/bsp-itub4/fi-apmscw'], #5
                ['br_renner','https://www.msn.com/en-us/money/stockdetails/history/bsp-lren3/fi-apmvxm'], #6
                ['br_gol','https://www.msn.com/en-us/money/stockdetails/history/bsp-goll4/fi-apmoa2'], #7
                ['br_magazineluiza','https://www.msn.com/en-us/money/stockdetails/history/bsp-mglu3/fi-apmxnm'], #8
                ['us_abercrombie','https://www.msn.com/en-us/money/stockdetails/history/nys-anf/fi-a1nikr'], #10
                ['us_boeing','https://www.msn.com/en-us/money/stockdetails/history/nys-ba/fi-a1o4ec'], #11
                ['us_beyondmeat','https://www.msn.com/en-us/money/stockdetails/history/nas-bynd/fi-bpocw7'], #12
                ['us_morganstanley','https://www.msn.com/en-us/money/stockdetails/history/nys-ms/fi-a1xyjc'], #13
                ['us_jpmorgan', 'https://www.msn.com/en-us/money/stockdetails/history/nys-jpm/fi-a1waa2'], #14
                ['us_exxonmobil','https://www.msn.com/en-us/money/stockdetails/history/nys-xom/fi-a269ec'], #15
                ['us_americanair','https://www.msn.com/en-us/money/stockdetails/history/nas-aal/fi-a1moa2'], #16
                ['us_cocacola','https://www.msn.com/en-us/money/stockdetails/history/nys-ko/fi-a1wljc'], #17
                ['us_tesla','https://www.msn.com/en-us/money/stockdetails/history/nas-tsla/fi-a24kar'], #18
                ['us_wsj','https://www.msn.com/en-us/money/stockdetails/history/nas-nws/fi-a1yx6h']] #19

driver = webdriver.Chrome(executable_path=r'C:\Users\drr19\Desktop\chromedriver.exe')

for company in my_companies:
    StockTracker_15(company, driver)
